+++ test/tests_basic_rendering.jl
@@ -18,7 +18,7 @@ const content2 = "efgh"
   @testset "Empty string force parse" begin
     r = html("", forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body></body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body></body></html>"
   end;
 
 
@@ -31,7 +31,7 @@ const content2 = "efgh"
   @testset "String no spaces force parse" begin
     r = html(content, forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>abcd</body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>abcd</body></html>"
   end;
 
 
@@ -44,7 +44,7 @@ const content2 = "efgh"
   @testset "String with 2 spaces force parse" begin
     r = html("  $content  ", forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>abcd  </body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>abcd  </body></html>"
   end;
 
 
@@ -57,7 +57,7 @@ const content2 = "efgh"
   @testset "String with &nbsp; force parse" begin
     r = html("  &nbsp;&nbsp;  ", forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>&nbsp;&nbsp;  </body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>&nbsp;&nbsp;  </body></html>"
   end;
 
 
@@ -70,7 +70,7 @@ const content2 = "efgh"
   @testset "String with 2 &nbsp; and 2 spaces force parse" begin
     r = html("&nbsp;&nbsp;$content  ", forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>&nbsp;&nbsp;abcd  </body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>&nbsp;&nbsp;abcd  </body></html>"
   end;
 
 
@@ -83,7 +83,7 @@ const content2 = "efgh"
   @testset "String with newline force parse" begin
     r = html("$content  \n  &nbsp;&nbsp;$content2", forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>abcd  \n&nbsp;&nbsp;efgh</body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>abcd  \n&nbsp;&nbsp;efgh</body></html>"
   end;
 
 
@@ -96,7 +96,7 @@ const content2 = "efgh"
   @testset "String with quotes force parse" begin
     r = html("He said \"wow!\"", forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>He said \"wow!\"\0</body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>He said \"wow!\"\0</body></html>"
   end;
 
 
@@ -109,7 +109,7 @@ const content2 = "efgh"
   @testset "String with quotes force parse" begin
     r = html(""" "" """, forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>\"\" </body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>\"\" </body></html>"
   end;
 
 
@@ -122,7 +122,7 @@ const content2 = "efgh"
   @testset "String with quotes force parse" begin
     r = html("\"\"", forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>\"\"\0</body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>\"\"\0</body></html>"
   end;
 
 
@@ -135,7 +135,7 @@ const content2 = "efgh"
   @testset "String with interpolated vars force parse" begin
     r = html("$(reverse(content))", forceparse = true)
 
-    @test String(r.body) == "<html><head></head><body>dcba</body></html>"
+    @test String(r.body) == "<!DOCTYPE html><html><head></head><body>dcba</body></html>"
   end;